requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langgraph 
openai
pymongo
python-dotenv
streamlit
pinecone
langchain_pinecone
langchain_core
langchain_text_splitters
pypdf
